Avoid using 'use ok' in test
RT #128679

ok.pm is not provided with old versions of Test::Simple.
This was introduced in 1.001010.

Rather than requesting a specific version of Test::Simple
be friendly and let the test passes even when using an outdated
version of Test::Simple.

'use ok' does not provide much more value, the test would fail just
after if we cannot load the module.